+++ frontend/app/src/components/history/events/forms/OnlineHistoryEventForm.vue
@@ -174,7 +174,7 @@ async function save(): Promise<boolean> {
     },
     location: get(location),
     locationLabel: get(locationLabel) || null,
-    notes: usedNotes || undefined,
+    notes: usedNotes ? usedNotes.trim() : undefined,
   };
 
   return await saveHistoryEventHandler(
@@ -300,7 +300,7 @@ const locationLabelSuggestions = computed(() =>
     <RuiDivider class="mb-6 mt-2" />
 
     <RuiTextArea
-      v-model.trim="notes"
+      v-model="notes"
       prepend-icon="sticky-note-line"
       data-cy="notes"
       variant="outlined"